Create docs
Created several different folders because it was confusing to me that the repository title we put up right from the start had "week01" as part of the  title. My apologies. Here is a plain docs folder. All the files here are duplicates of the other files saved on the drive at the time the repository URL was submitted